cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "enclaiv"
version = "0.1.0"
description = "CLI for the Enclaiv AI agent sandboxing platform — runs agents in hardware-isolated Unikraft unikernel VMs."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "httpx>=0.27",
]

[project.scripts]
enclaiv = "enclaiv.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["enclaiv*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
